+++ CHANGES.md
@@ -26,6 +26,7 @@ CUPS v2.4rc1 (Pending)
 - Removed support for the (long deprecated and unused) `FontPath`,
   `LPDConfigFile`, `RIPCache`, and `SMBConfigFile` directives in `cupsd.conf`
   and `cups-files.conf`.
+- Add retry in IPP backend if Validate-Job failed
 
 
 CUPS v2.3.3op2 (February 1, 2021)